+++ babel.config.js
@@ -1,6 +1,9 @@
-const devPresets = ['@vue/cli-plugin-babel/preset'];
-const buildPresets = [['@babel/preset-env', {}]];
-
+// Babel config for storybook
 module.exports = {
-  presets: process.env.NODE_ENV === 'development' ? devPresets : buildPresets,
+  presets: [['@babel/preset-env', {}]],
+  plugins: [
+    ['@babel/plugin-proposal-private-property-in-object', { loose: true }],
+    ['@babel/plugin-proposal-private-methods', { loose: true }],
+    ['@babel/plugin-proposal-class-properties', { loose: true }],
+  ],
 };