+++ src/utils/utils.js
@@ -1,0 +1,7 @@
+export const getParentByName = (parent, name) => {
+  while (parent) {
+    if (parent.proxy.$options.name === name) return parent;
+    parent = parent.parent;
+  }
+  return false;
+};